+++ src/app/products/services/product.service.spec.ts
@@ -13,19 +13,19 @@ describe('ProductService Tests', () => {
         const service = new ProductService();
         service.getProducts()
             .subscribe(x => {
-                expect(x.length).toBeGreaterThanOrEqual(3)
+                expect(x.length).toBeGreaterThanOrEqual(3);
                 done();
             });
     });
 
     it('Service should update product', () => {
-        var service = TestBed.get(ProductService) as ProductService;
+        const service = TestBed.get(ProductService) as ProductService;
 
-        var product = service.getProduct(3) as IProduct;
-        product.name = "Updated Product"
+        let product = service.getProduct(3) as IProduct;
+        product.name = 'Updated Product';
         service.updateProduct(product);
 
         product = service.getProduct(3) as IProduct;
-        expect(product.name).toBe("Updated Product");
+        expect(product.name).toBe('Updated Product');
     });
 });