+++ src/app/shared/pipes/order-by.pipe.spec.ts
@@ -3,13 +3,13 @@ import { OrderByPipe } from './order-by.pipe';
 
 function getObjectsArray(): any[] {
     return [ {
-        name: "B",
+        name: 'B',
         age: 3
     }, {
-        name: "C",
+        name: 'C',
         age: 1
     }, {
-        name: "A",
+        name: 'A',
         age: 2
     }];
 }
@@ -19,8 +19,8 @@ describe('OrderByPipe Tests', () => {
     const pipe = new OrderByPipe();
 
     it('Asc order array of objects by "name" property', () => {
-        var items = getObjectsArray();
-        var orderedItems = pipe.transform([...items], 'name', true);
+        const items = getObjectsArray();
+        const orderedItems = pipe.transform([...items], 'name', true);
 
         expect(orderedItems[0]).toBe(items[2]);
         expect(orderedItems[1]).toBe(items[0]);
@@ -28,8 +28,8 @@ describe('OrderByPipe Tests', () => {
     });
 
     it('Desc order array of objects by "age" property', () => {
-        var items = getObjectsArray();
-        var orderedItems = pipe.transform([...items], 'age', false);
+        const items = getObjectsArray();
+        const orderedItems = pipe.transform([...items], 'age', false);
 
         expect(orderedItems[0]).toBe(items[0]);
         expect(orderedItems[1]).toBe(items[2]);